rename make api request
