+++ frontend/app/.server/domain/services/language-referral-type-service-default.ts
@@ -3,7 +3,7 @@ import { Ok, Err } from 'oxide.ts';
 
 import type { LanguageReferralType, LocalizedLanguageReferralType } from '~/.server/domain/models';
 import type { LanguageReferralTypeService } from '~/.server/domain/services/language-referral-type-service';
-import { apiFetch } from '~/.server/domain/services/makeApiRequest';
+import { apiFetch } from '~/.server/domain/services/make-api-request';
 import { AppError } from '~/errors/app-error';
 import { ErrorCodes } from '~/errors/error-codes';
 import { HttpStatusCodes } from '~/errors/http-status-codes';
